+++ README.md
@@ -1,65 +1,82 @@
-# Pull Request Analysis with MCP
+# MCP GitHub Operations PoC
 
-## Prerequisites
+This project provides a modular, extensible framework for automating GitHub repository, PR, CI, and documentation operations, with integration to Notion and Confluence.
 
-1. Claude Desktop
+## Features
+- **Pull Request Tools**: Fetch, create, and merge PRs; create Notion and Confluence pages for PR analysis.
+- **Repository Tools**: Create repositories, set up language templates, generate Dockerfiles and README.md files.
+- **CI Tools**: Analyze GitHub Actions pipeline results and provide recommendations.
+- **Test Generation**: Auto-generate Python unittest stubs for all functions/classes in the repo.
 
-2. Confluence Account
+## Modular Architecture
 
-3. Notion Account
+The codebase is organized into logical modules for maintainability:
 
-4. Python
+```
+project-root/
+  github_ops/
+    auth.py         # Authentication and client setup (GitHub, Notion, Confluence)
+    pr_tools.py     # Pull request tools (fetch, create, merge, Notion/Confluence pages)
+    repo_tools.py   # Repository management tools (create repo, templates, Dockerfile, README)
+    ci_tools.py     # CI pipeline analysis tools
+    testgen.py      # Test generation tools
+  pr_analyzer.py    # Main orchestration/entry point
+  tests/            # Test stubs auto-generated here
+  requirements.txt  # Python dependencies
+  README.md         # This file
+```
 
-Project leverages MCP to:
+### Tool Registration
+Each tool group is registered in `pr_analyzer.py`:
+```python
+from github_ops.pr_tools import register_pr_tools
+from github_ops.repo_tools import register_repo_tools
+from github_ops.ci_tools import register_ci_tools
+from github_ops.testgen import register_testgen_tools
 
- 1. Analyze PRs prior to merges, making the code review process easier. The results of the analysis are published to Notion and Confluence Pages.
+# ...
+register_pr_tools(...)
+register_repo_tools(...)
+register_ci_tools(...)
+register_testgen_tools(...)
+```
 
- 2. Create a github repo
+## Extending the System
+- To add new tools, create a new module in `github_ops/` or extend an existing one.
+- Register new tools using the `@mcp.tool()` decorator in the appropriate module.
+- Import and call the registration function in `pr_analyzer.py`.
 
- 3. Create a PR
+## Setup & Usage
+1. **Install dependencies:**
+   ```bash
+   pip install -r requirements.txt
+   ```
+2. **Set environment variables:** (see `.env.copy` for required keys)
+3. **Run the server:**
+   ```bash
+   python pr_analyzer.py
+   ```
 
- 4. Merge a PR
+## High-Level Architecture Diagram
 
-As part of the analysis, the model also recommends improvements to each PR, further improving code quality.
+```
+flowchart TD
+    subgraph github_ops
+        A[auth.py]
+        B[pr_tools.py]
+        C[repo_tools.py]
+        D[ci_tools.py]
+        E[testgen.py]
+    end
+    F[pr_analyzer.py] -->|registers tools| B
+    F --> C
+    F --> D
+    F --> E
+    F --> A
+    B -->|uses| A
+    C -->|uses| A
+    D -->|uses| A
+```
 
-Link to Notion page: https://www.notion.so/PR-Summaries-1f325dacde7e8015ba2dee4e9c8976ed
-
-### Test Project
-
-Create and populate your .env file with variables defined in .env.copy
-
-Create virtual environment:
-
-`python3 -m venv venv`
-
-Activate virtual environment:
-
-`source venv/bin/activate`
-
-Install dependencies:
-
-`pip install -r requirements.txt`
-
-Run script/mcp tools:
-
-`python pr_analyzer.py`
-
-
-Launch claude desktop and test with prompt below:
-
-`Analyze this github pull request: https://github.com/o-ouma/mcp-poc/pull/6`
-
-After analysis claude publishes the PR analysis summary to notion and confluence pages
-
-
-### To Do:
-
-1. Add create repo feature
-
-2. Add repo templates for python, node, angular, golang etc.
-
-3. Add feature to create PR
-
-4. Merge PR feature
-
-5. Analyze pipeline results
+## License
+MIT
